+++ e2e/e2etests/e2etests.go
@@ -294,7 +294,7 @@ var (
 	DepdencyAllBitcoinDeposits = runner.NewE2EDependency("all_bitcoin_deposits")
 )
 
-const V2ZETATestVersion = "v33.0.0"
+const V2ZetaVersion = "v33.0.0"
 
 // AllE2ETests is an ordered list of all e2e tests
 var AllE2ETests = []runner.E2ETest{
@@ -308,7 +308,7 @@ var AllE2ETests = []runner.E2ETest{
 			{Description: "amount in wei", DefaultValue: "1000000000000000000"},
 		},
 		TestZetaDeposit,
-		runner.WithMinimumVersion(V2ZETATestVersion),
+		runner.WithMinimumVersion(V2ZetaVersion),
 	),
 	runner.NewE2ETest(
 		TestZetaDepositAndCallName,
@@ -317,7 +317,7 @@ var AllE2ETests = []runner.E2ETest{
 			{Description: "amount", DefaultValue: "100000000000000000000"},
 		},
 		TestZetaDepositAndCall,
-		runner.WithMinimumVersion(V2ZETATestVersion),
+		runner.WithMinimumVersion(V2ZetaVersion),
 	),
 	runner.NewE2ETest(
 		TestZetaDepositAndCallRevertName,
@@ -326,14 +326,14 @@ var AllE2ETests = []runner.E2ETest{
 			{Description: "amount", DefaultValue: "10000000000000000000"},
 		},
 		TestZetaDepositAndCallRevert,
-		runner.WithMinimumVersion(V2ZETATestVersion),
+		runner.WithMinimumVersion(V2ZetaVersion),
 	),
 	runner.NewE2ETest(
 		TestZetaDepositRevertAndAbortName,
 		"deposit Zeta into ZEVM, revert, then abort with onAbort because revert fee cannot be paid",
 		[]runner.ArgDefinition{},
 		TestZetaDepositRevertAndAbort,
-		runner.WithMinimumVersion("v32.0.0"),
+		runner.WithMinimumVersion(V2ZetaVersion),
 	),
 	runner.NewE2ETest(
 		TestZetaDepositAndCallRevertWithCallName,
@@ -342,7 +342,7 @@ var AllE2ETests = []runner.E2ETest{
 			{Description: "amount", DefaultValue: "10000000000000000000"},
 		},
 		TestZetaDepositAndCallRevertWithCall,
-		runner.WithMinimumVersion(V2ZETATestVersion),
+		runner.WithMinimumVersion(V2ZetaVersion),
 	),
 	runner.NewE2ETest(
 		TestZetaDepositAndCallNoMessageName,
@@ -351,7 +351,7 @@ var AllE2ETests = []runner.E2ETest{
 			{Description: "amount", DefaultValue: "10000000000000000000"},
 		},
 		TestZetaDepositAndCallNoMessage,
-		runner.WithMinimumVersion(V2ZETATestVersion),
+		runner.WithMinimumVersion(V2ZetaVersion),
 	),
 	runner.NewE2ETest(
 		TestZetaWithdrawName,
@@ -360,7 +360,7 @@ var AllE2ETests = []runner.E2ETest{
 			{Description: "amount", DefaultValue: "1000"},
 		},
 		TestZetaWithdraw,
-		runner.WithMinimumVersion(V2ZETATestVersion),
+		runner.WithMinimumVersion(V2ZetaVersion),
 	),
 	runner.NewE2ETest(
 		TestZetaWithdrawAndCallName,
@@ -370,7 +370,7 @@ var AllE2ETests = []runner.E2ETest{
 			{Description: "gas limit for withdraw and call", DefaultValue: "250000"},
 		},
 		TestZetaWithdrawAndCall,
-		runner.WithMinimumVersion(V2ZETATestVersion),
+		runner.WithMinimumVersion(V2ZetaVersion),
 	),
 	runner.NewE2ETest(
 		TestZetaWithdrawAndCallRevertName,
@@ -379,7 +379,7 @@ var AllE2ETests = []runner.E2ETest{
 			{Description: "amount", DefaultValue: "1000"},
 		},
 		TestZetaWithdrawAndCallRevert,
-		runner.WithMinimumVersion(V2ZETATestVersion),
+		runner.WithMinimumVersion(V2ZetaVersion),
 	),
 	runner.NewE2ETest(
 		TestZetaWithdrawAndCallRevertWithCallName,
@@ -388,7 +388,7 @@ var AllE2ETests = []runner.E2ETest{
 			{Description: "amount", DefaultValue: "1000"},
 		},
 		TestZetaWithdrawAndCallRevertWithCall,
-		runner.WithMinimumVersion(V2ZETATestVersion),
+		runner.WithMinimumVersion(V2ZetaVersion),
 	),
 	runner.NewE2ETest(
 		TestZetaWithdrawRevertAndAbortName,
@@ -398,7 +398,7 @@ var AllE2ETests = []runner.E2ETest{
 			{Description: "gas limit for withdraw and call", DefaultValue: "250000"},
 		},
 		TestZetaWithdrawRevertAndAbort,
-		runner.WithMinimumVersion(V2ZETATestVersion),
+		runner.WithMinimumVersion(V2ZetaVersion),
 	),
 	runner.NewE2ETest(
 		TestZetaWithdrawAndArbitraryCallName,
@@ -407,7 +407,7 @@ var AllE2ETests = []runner.E2ETest{
 			{Description: "amount", DefaultValue: "1000"},
 		},
 		TestZetaWithdrawAndArbitraryCall,
-		runner.WithMinimumVersion(V2ZETATestVersion),
+		runner.WithMinimumVersion(V2ZetaVersion),
 	),
 	runner.NewE2ETest(
 		TestZetaWithdrawAndCallNoMessageName,
@@ -417,7 +417,7 @@ var AllE2ETests = []runner.E2ETest{
 			{Description: "gas limit for withdraw and call", DefaultValue: "250000"},
 		},
 		TestZetaWithdrawAndCallNoMessage,
-		runner.WithMinimumVersion(V2ZETATestVersion),
+		runner.WithMinimumVersion(V2ZetaVersion),
 	),
 	runner.NewE2ETest(
 		TestETHDepositName,
@@ -1103,7 +1103,7 @@ var AllE2ETests = []runner.E2ETest{
 			{Description: "amount in mist", DefaultValue: "1000000"},
 		},
 		TestSuiWithdraw,
-		runner.WithMinimumVersion("v29.0.0"),
+		runner.WithMinimumVersion("v33.0.0"),
 	),
 	runner.NewE2ETest(
 		TestSuiWithdrawAndCallName,
@@ -1113,7 +1113,7 @@ var AllE2ETests = []runner.E2ETest{
 			{Description: "gas limit for withdraw and call", DefaultValue: "100000"},
 		},
 		TestSuiWithdrawAndCall,
-		runner.WithMinimumVersion("v32.0.0"),
+		runner.WithMinimumVersion("v33.0.0"),
 	),
 	runner.NewE2ETest(
 		TestSuiWithdrawRevertWithCallName,
@@ -1122,7 +1122,7 @@ var AllE2ETests = []runner.E2ETest{
 			{Description: "amount in mist", DefaultValue: "1000000"},
 		},
 		TestSuiWithdrawRevertWithCall,
-		runner.WithMinimumVersion("v30.0.0"),
+		runner.WithMinimumVersion("v33.0.0"),
 	),
 	runner.NewE2ETest(
 		TestSuiWithdrawAndCallRevertWithCallName,
@@ -1132,7 +1132,7 @@ var AllE2ETests = []runner.E2ETest{
 			{Description: "gas limit for withdraw and call", DefaultValue: "100000"},
 		},
 		TestSuiWithdrawAndCallRevertWithCall,
-		runner.WithMinimumVersion("v32.0.0"),
+		runner.WithMinimumVersion("v33.0.0"),
 	),
 	runner.NewE2ETest(
 		TestSuiTokenWithdrawName,
@@ -1141,7 +1141,7 @@ var AllE2ETests = []runner.E2ETest{
 			{Description: "amount in base unit", DefaultValue: "100000"},
 		},
 		TestSuiTokenWithdraw,
-		runner.WithMinimumVersion("v29.0.0"),
+		runner.WithMinimumVersion("v33.0.0"),
 	),
 	runner.NewE2ETest(
 		TestSuiTokenWithdrawAndCallName,
@@ -1151,7 +1151,7 @@ var AllE2ETests = []runner.E2ETest{
 			{Description: "gas limit for withdraw and call", DefaultValue: "100000"},
 		},
 		TestSuiTokenWithdrawAndCall,
-		runner.WithMinimumVersion("v32.0.0"),
+		runner.WithMinimumVersion("v33.0.0"),
 	),
 	runner.NewE2ETest(
 		TestSuiTokenWithdrawAndCallRevertWithCallName,
@@ -1161,7 +1161,7 @@ var AllE2ETests = []runner.E2ETest{
 			{Description: "gas limit for withdraw and call", DefaultValue: "100000"},
 		},
 		TestSuiTokenWithdrawAndCallRevertWithCall,
-		runner.WithMinimumVersion("v32.0.0"),
+		runner.WithMinimumVersion("v33.0.0"),
 	),
 	runner.NewE2ETest(
 		TestSuiDepositRestrictedName,
@@ -1179,7 +1179,7 @@ var AllE2ETests = []runner.E2ETest{
 			{Description: "amount in mist", DefaultValue: "1000000"},
 		},
 		TestSuiWithdrawRestrictedAddress,
-		runner.WithMinimumVersion("v30.0.0"),
+		runner.WithMinimumVersion("v33.0.0"),
 	),
 	runner.NewE2ETest(
 		TestSuiWithdrawInvalidReceiverName,
@@ -1189,6 +1189,7 @@ var AllE2ETests = []runner.E2ETest{
 			{Description: "amount in mist", DefaultValue: "1000000"},
 		},
 		TestSuiWithdrawInvalidReceiver,
+		runner.WithMinimumVersion("v33.0.0"),
 	),
 	/*
 	 Bitcoin tests