+++ e2e/runner/accounting.go
@@ -1,18 +1,17 @@
 package runner
 
 import (
-	"encoding/json"
 	"fmt"
-	"io"
 	"math/big"
-	"net/http"
 
+	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/gagliardetto/solana-go/rpc"
 	"github.com/stretchr/testify/require"
 
 	zetacrypto "github.com/zeta-chain/node/pkg/crypto"
+	fungibletypes "github.com/zeta-chain/node/x/fungible/types"
 	observertypes "github.com/zeta-chain/node/x/observer/types"
 )
 
@@ -171,6 +170,39 @@ func (r *E2ERunner) CheckSolanaTSSBalance() error {
 	return nil
 }
 
+// CheckSUITSSBalance checks the TSS balance on Sui against the ZRC20 total supply
+func (r *E2ERunner) CheckSUITSSBalance() error {
+	gatewayBalance, err := r.SuiGetGatewaySUIBalance()
+	if err != nil {
+		return fmt.Errorf("failed to get SUI balance for Sui gateway: %w", err)
+	}
+
+	zrc20Supply, err := r.SUIZRC20.TotalSupply(&bind.CallOpts{})
+	if err != nil {
+		return err
+	}
+
+	// subtract value from the gas stability pool because of the artificial minting bug
+	// TODO: remove on the chain upgrade to v33
+	// https://github.com/zeta-chain/node/issues/4034
+	gasStabiltiyPoolBalance, err := r.SUIZRC20.BalanceOf(&bind.CallOpts{}, fungibletypes.GasStabilityPoolAddressEVM())
+	if err != nil {
+		return fmt.Errorf("failed to get SUI gas stability pool balance: %w", err)
+	}
+	zrc20Supply = zrc20Supply.Sub(zrc20Supply, gasStabiltiyPoolBalance)
+
+	// Subtract 0.1 SUI to take in consideration the 0.1 SUI minted in the gas pool
+	// TODO: a proper implementation is to implement a donate method in Sui Contract and use it to donate 0.1 SUI
+	// https://github.com/zeta-chain/protocol-contracts-sui/issues/58
+	zrc20Supply = zrc20Supply.Sub(zrc20Supply, big.NewInt(100000000))
+
+	if gatewayBalance.Cmp(zrc20Supply) < 0 {
+		return fmt.Errorf("SUI: TSS balance (%d) < ZRC20 TotalSupply (%d) ", gatewayBalance, zrc20Supply)
+	}
+	r.Logger.Info("SUI: TSS balance (%d) >= ZRC20 TotalSupply (%d)", gatewayBalance, zrc20Supply)
+	return nil
+}
+
 func (r *E2ERunner) checkERC20TSSBalance() error {
 	custodyBalance, err := r.ERC20.BalanceOf(&bind.CallOpts{}, r.ERC20CustodyAddr)
 	if err != nil {
@@ -189,6 +221,7 @@ func (r *E2ERunner) checkERC20TSSBalance() error {
 }
 
 func (r *E2ERunner) checkZetaTSSBalance() {
+	// get total ZETA locked
 	zetaLockedLegacyConnector, err := r.ZetaEth.BalanceOf(&bind.CallOpts{}, r.ConnectorEthAddr)
 	require.NoError(r, err, "BalanceOf failed for legacy connector")
 
@@ -197,15 +230,14 @@ func (r *E2ERunner) checkZetaTSSBalance() {
 
 	zetaLocked := big.NewInt(0).Add(zetaLockedLegacyConnector, zetaLockedConnectorNative)
 
-	resp, err := http.Get("http://zetacore0:1317/cosmos/bank/v1beta1/supply/by_denom?denom=azeta")
-	require.NoError(r, err)
-	defer resp.Body.Close()
-	body, err := io.ReadAll(resp.Body)
-	require.NoError(r, err)
-	var result Response
-	err = json.Unmarshal(body, &result)
-	require.NoError(r, err)
-	zetaSupply, _ := big.NewInt(0).SetString(result.Amount.Amount, 10)
+	// get ZETA supply
+	res, err := r.Clients.Zetacore.Bank.SupplyOf(r.Ctx, &banktypes.QuerySupplyOfRequest{
+		Denom: "azeta",
+	})
+	require.NoError(r, err, "SupplyOf failed for azeta")
+	require.NotNil(r, res, "SupplyOf response is nil")
+	zetaSupply := res.Amount.Amount.BigInt()
+
 	if zetaLocked.Cmp(zetaSupply) < 0 {
 		r.Logger.Info("ZETA: TSS balance (%d) < ZRC20 TotalSupply (%d)", zetaLocked, zetaSupply)
 	} else {